+++ CHANGELOG.md
@@ -1,3 +1,4 @@
 Changes:
 * Render global 7tv emotes
+* Show tooltip with emote's name when hovering over it